Change work_target_raw_azimuth to float type
Changed work_target_raw_azimuth from int to float for resolution finer than 1 degree, DL8EBL , Olaf , Was already identified by G8RWG see (at bottom of article) 

https://g8rwg.uk/articles/kn3g_rotator_controller_bench_tests/

+++ k3ng_rotator_controller/k3ng_rotator_controller.ino
@@ -11960,7 +11960,7 @@ void service_request_queue(){
     control_port->flush();
   #endif // DEBUG_LOOP
 
-  int work_target_raw_azimuth = 0;
+  float work_target_raw_azimuth = 0; // changed from "int" to "float" to work with resolutions finer then one degree, e.g. 0.1 degree
   byte direction_to_go = 0;
   byte within_tolerance_flag = 0;
